Bugfix in lights script to support device without GPIO

+++ system/bin/lights.sh
@@ -12,16 +12,18 @@ init() {
        direction=`cat /sys/class/gpio/gpio41_pd7/direction`
 
        if [ "$direction" != "out" ]; then
-       	echo 41 > /sys/class/gpio/export 2>/dev/null
-       	echo 42 > /sys/class/gpio/export 2>/dev/null
-       	echo 43 > /sys/class/gpio/export 2>/dev/null
-       	echo out > /sys/class/gpio/gpio41_pd7/direction 2>/dev/null
-       	echo out > /sys/class/gpio/gpio42_pd6/direction 2>/dev/null
-       	echo out > /sys/class/gpio/gpio43_pd5/direction 2>/dev/null
+       	echo 41 > /sys/class/gpio/export
+       	echo 42 > /sys/class/gpio/export
+       	echo 43 > /sys/class/gpio/export
+       	echo out > /sys/class/gpio/gpio41_pd7/direction
+       	echo out > /sys/class/gpio/gpio42_pd6/direction
+       	echo out > /sys/class/gpio/gpio43_pd5/direction
        fi
 }
 
-init
+if [ -f $lamps["GREEN"] ]; then
+	init
+fi
 
 stopblink()
 {
@@ -29,29 +31,31 @@ stopblink()
 }
 
 
-# PARSE INPUT DATA
-if [ "$2" = "ON" ]; then
-  	stopblink
-	echo 1 > ${lamps[$1]}
-elif [ "$2" = "OFF" ]; then
-	stopblink
-	echo 0 > ${lamps[$1]}
-elif [ "$2" = "BLINK" ]; then
-	stopblink
-	if [ ! -f $blinkfile ]; then
-		touch $blinkfile
-		while test -f $blinkfile
-		do
-			if [ -f $blinkfile ]; then
-				echo 1 > ${lamps["$1"]}
-			fi
-			sleep 0.6
-			if [ -f $blinkfile ]; then
-				echo 0 > ${lamps["$1"]}
-			fi
-			sleep 0.6
-		done
+if [ -f ${lamps[$1]} ]; then
+	# PARSE INPUT DATA
+	if [ "$2" = "ON" ]; then
+	  	stopblink
+		echo 1 > ${lamps[$1]}
+	elif [ "$2" = "OFF" ]; then
+		stopblink
+		echo 0 > ${lamps[$1]}
+	elif [ "$2" = "BLINK" ]; then
+		stopblink
+		if [ ! -f $blinkfile ]; then
+			touch $blinkfile
+			while test -f $blinkfile
+			do
+				if [ -f $blinkfile ]; then
+					echo 1 > ${lamps["$1"]}
+				fi
+				sleep 0.6
+				if [ -f $blinkfile ]; then
+					echo 0 > ${lamps["$1"]}
+				fi
+				sleep 0.6
+			done
+		fi
+	elif [ "$2" = "STOPBLINK" ]; then
+		stopblink
 	fi
-elif [ "$2" = "STOPBLINK" ]; then
-	stopblink
 fi
